API: application controller restructuring. Update Application Type on Submission list
