+++ drr/src/API/EMCR.DRR/Model/Submission.cs
@@ -1,11 +1,13 @@
-﻿using EMCR.DRR.Controllers;
+﻿using System.ComponentModel;
+using System.Text.Json.Serialization;
+using EMCR.DRR.Controllers;
 
 namespace EMCR.DRR.API.Model
 {
     public class Submission
     {
         public required string Id { get; set; }
-        public required string ApplicationType { get; set; }
+        public required ApplicationType ApplicationType { get; set; }
         public string? ExistingFpId { get; set; }
         public required string ProjectTitle { get; set; }
         public required SubmissionPortalStatus Status { get; set; }
@@ -14,5 +16,15 @@ namespace EMCR.DRR.API.Model
         public DateTime? SubmittedDate { get; set; }
         public required string[] PartneringProponents { get; set; }
         public FundingStream? FundingStream { get; set; }
+    }
+
+    [JsonConverter(typeof(JsonStringEnumConverter))]
+    public enum ApplicationType
+    {
+        [Description("EOI")]
+        EOI,
+
+        [Description("Full Proposal")]
+        FP,
     }
 }
